lambertian shading fixed

+++ rgbapixel.cpp
@@ -7,19 +7,19 @@
  */
 #include "rgbapixel.h"
 
-RGBAPixel::RGBAPixel() : red(255), green(255), blue(255), alpha(255)
+RGBAPixel::RGBAPixel() : red(255.0), green(255.0), blue(255.0), alpha(255.0)
 {
 	/* nothing */
 }
 
-RGBAPixel::RGBAPixel(uint8_t r, uint8_t g, uint8_t b)
-	: red(r), green(g), blue(b), alpha(255)
+RGBAPixel::RGBAPixel(double r, double g, double b)
+	: red(r), green(g), blue(b), alpha(255.0)
 {
 	/* nothing */
 }
 
-RGBAPixel::RGBAPixel(uint8_t r, uint8_t g, uint8_t b,
-		             uint8_t a) : red(r), green(g), blue(b), alpha(a)
+RGBAPixel::RGBAPixel(double r, double g, double b,
+		             double a) : red(r), green(g), blue(b), alpha(a)
 {
 	/* nothing */
 }
